+++ Runtime/Scripts/Proto/Message.cs
@@ -1,4 +1,6 @@
 
+using PJ.Native.PubSub;
+
 namespace PJ.Native.Proto
 {
     public partial class Message
@@ -12,6 +14,22 @@ namespace PJ.Native.Proto
         {
             Key = key;
             Container = container;
+        }
+    }
+
+    public partial class Envelope
+    {
+        public Envelope(Message message, int senderID)
+        {
+            Message = message;
+            SenderID = senderID;
+        }
+
+        public Envelope(Message message, int senderID, int receiverID)
+        {
+            Message = message;
+            SenderID = senderID;
+            ReceiverID = receiverID;
         }
     }
 }