+++ Runtime/Scripts/PubSub/Tag.cs
@@ -5,34 +5,38 @@ using Unity.Collections.LowLevel.Unsafe;
 using Unity.VisualScripting;
 using UnityEngine;
 
-namespace PJ.Native.Messenger
+namespace PJ.Native.PubSub
 {
     public partial struct Tag 
     {
 
         private string name;
         private ulong id;
+        private List<string> names;
         public string Name => name;
 
         private Tag(string name)
         {
             this.name = name;
+            this.names = null;
             this.id = TagGenerator.Generate();
         }
 
         private Tag(ulong id)
         {
             this.name = id.ToString();
+            this.names = null;
             this.id = id;
         }
 
         private Tag(string name, ulong id)
         {
             this.name = name;
+            this.names = null;
             this.id = id;
         }
 
-        public static Tag Create(string name)
+        public static Tag Named(string name)
         {
             if (namedTagDic.ContainsKey(name))
             {
@@ -43,6 +47,41 @@ namespace PJ.Native.Messenger
                 var tag = new Tag(name);
                 namedTagDic.Add(name, tag);
                 tagDic.Add(tag.id, tag);
+                return tag;
+            }
+        }
+        
+        private static Tag NamedWithID(string name, ulong id)
+        {
+            if(tagDic.ContainsKey(id))
+            {
+                Tag tag = tagDic[id];
+                tag.name = name;
+                return tag;
+            }
+            else
+            {
+                var tag = new Tag(name, id);
+                namedTagDic.Add(name, tag);
+                tagDic.Add(id, tag);
+                return tag;
+            }
+        }
+
+        public static Tag Named(IEnumerable<string> names)
+        {
+            ulong joined = 0b0;
+            foreach(var name in names)
+            {
+                Tag named = Named(name);
+                joined |= named.id;
+            }
+            if(tagDic.ContainsKey(joined))
+                return tagDic[joined];
+            else
+            {
+                Tag tag = new Tag(joined);
+                tagDic.Add(joined, tag);
                 return tag;
             }
         }
@@ -76,15 +115,60 @@ namespace PJ.Native.Messenger
                 return tag;
             }
         }
+
+        public Tag Unjoin(Tag tag2)
+        {
+            ulong unjoined = this.id & ~tag2.id;
+            if(tagDic.ContainsKey(unjoined))
+            {
+                return tagDic[unjoined];
+            }
+            else
+            {
+                var tag = new Tag(unjoined);
+                tagDic.Add(unjoined, tag);
+                return tag;
+            }
+        }
+
+        private bool Contains(ulong id)
+        {
+            return (this.id & id) == id;
+        }
+
+        public Tag Join(Tag tag)
+        {
+            return Tag.Join(this, tag);
+        }
         
         public bool Contains(Tag tag)
         {
-            return (id & tag.id) == tag.id;
+            return Contains(tag.id);
+            // return (id & tag.id) == tag.id;
         }
 
         public bool Except(Tag tag)
         {
             return (id & tag.id) == 0;
+        }
+
+        public List<string> Names
+        {
+            get
+            {
+                if(names != null)
+                    return names;
+
+                names = new List<string>();
+                for(ulong findID = 0b1; findID != 0b0; findID <<= 1)
+                {
+                    if(this.Contains(findID) && tagDic.TryGetValue(findID, out Tag findTag))
+                    {
+                        names.Add(name);
+                    }
+                }
+                return names;
+            }
         }
     }
 
@@ -115,9 +199,9 @@ namespace PJ.Native.Messenger
 
     public partial struct Tag
     {
-        public static Tag None = new Tag("None", 0b0);
-        public static Tag Native = new Tag("Native");
-        public static Tag Game = new Tag("Game");
+        public static Tag None = Tag.NamedWithID("None", 0b0);
+        public static Tag Native = Tag.Named("Native");
+        public static Tag Game = Tag.Named("Game");
         
     }
 }